+++ test/helper.rb
@@ -1,6 +1,6 @@
 require 'simplecov'
 SimpleCov.start do
-  add_filter '/test|deps|method_decorators/'
+  add_filter '/test|deps/'
 end
 
 require 'test/unit'